Add file share

+++ WebSite/app/helpers/Media.php
@@ -1,44 +1,85 @@
 <?php
+
 declare(strict_types=1);
 
 namespace app\helpers;
+
+use app\models\DataHelper;
+use app\models\SharedFileDataHelper;
 
 class Media
 {
     private const MEDIA_PATH =  __DIR__ . '/../../data/media/';
 
-    public function __construct()
+    public function __construct(private DataHelper $dataHelper, private SharedFileDataHelper $sharedFileDataHelper)
     {
         if (!file_exists(self::MEDIA_PATH)) mkdir(self::MEDIA_PATH, 0755, true);
     }
 
-    public function deleteFile(int $year, int $month, string $filename)
+    public function deleteFile(int $year, int $month, string $filename): array
     {
         $filename = basename($filename);
-        $filePath = self::MEDIA_PATH . $year . '/' . $month . '/' . $filename;
+        $filePath = self::MEDIA_PATH . $year . DIRECTORY_SEPARATOR . sprintf("%02d", $month) . DIRECTORY_SEPARATOR . $filename;
         $response = ['success' => false, 'message' => ''];
-
         if (!file_exists($filePath)) $response['message'] = 'Fichier non trouvé';
         else {
             if (unlink($filePath)) {
                 $response['success'] = true;
                 $response['message'] = 'Fichier supprimé avec succès';
 
-                $monthDir = self::MEDIA_PATH . $year . '/' . $month;
+                $monthDir = self::MEDIA_PATH . $year . DIRECTORY_SEPARATOR . sprintf("%02d", $month);
                 if (count(glob("$monthDir/*")) === 0) {
                     rmdir($monthDir);
 
                     $yearDir = self::MEDIA_PATH . $year;
-                    if (count(glob("$yearDir/*")) === 0) {
-                        rmdir($yearDir);
-                    }
+                    if (count(glob("$yearDir/*")) === 0) rmdir($yearDir);
                 }
             } else $response['message'] = 'Erreur lors de la suppression du fichier';
         }
         return $response;
     }
 
-    public function uploadFile($file)
+    public function getSharefile(int $year, int $month, string $filename): array | false
+    {
+        $filePath = self::MEDIA_PATH . $year . DIRECTORY_SEPARATOR . sprintf("%02d", $month) . DIRECTORY_SEPARATOR . $filename;
+        $sharedFile = $this->dataHelper->get('SharedFile', ['Item' => $filePath], 'IdGroup, OnlyForMembers');
+        return ['success' => $sharedFile !== false, 'data' => $sharedFile];
+    }
+
+    public static function getMediaPath(): string
+    {
+        return self::MEDIA_PATH;
+    }
+
+    public function isShared(string $filePath): array
+    {
+        $sharedFile = $this->sharedFileDataHelper->getSharedFile($filePath);
+        return ['success' => $sharedFile !== false && $sharedFile->Token !== null, 'data' => $sharedFile];
+    }
+
+    public function removeFileShare(int $year, int $month, string $filename): array
+    {
+        $filePath = self::MEDIA_PATH . $year . DIRECTORY_SEPARATOR . sprintf("%02d", $month) . DIRECTORY_SEPARATOR . $filename;
+        $sharedFile = $this->dataHelper->get('SharedFile', ['Item' => $filePath], 'Id');
+        if ($sharedFile) $this->dataHelper->set('SharedFile', ['Token' => null], ['Id' => $sharedFile->Id]);
+        return ['success' => true, 'message' => ''];
+    }
+
+    public function sharefile(int $year, int $month, string $filename, ?int $idGroup, int $onlyForMembers): array
+    {
+        $filePath = self::MEDIA_PATH . $year . DIRECTORY_SEPARATOR . sprintf("%02d", $month) . DIRECTORY_SEPARATOR . $filename;
+        $sharedFile = $this->dataHelper->get('SharedFile', ['Item' => $filePath], 'Id, Token');
+        $newToken = bin2hex(random_bytes(32));
+        $this->dataHelper->set('SharedFile', [
+            'Item' => $filePath,
+            'IdGroup' => $idGroup,
+            'OnlyForMembers' => $onlyForMembers,
+            'Token' => $sharedFile != false ? $sharedFile->Token ?? $newToken : $newToken
+        ],  $sharedFile != false ? ['Id' => $sharedFile->Id] : []);
+        return ['success' => true, 'message' => ''];
+    }
+
+    public function uploadFile($file): array
     {
         $year = date('Y');
         $month = date('m');
@@ -69,10 +110,5 @@ class Media
             ];
         } else $response['message'] = 'Erreur lors de l\'enregistrement du fichier';
         return $response;
-    }
-
-    public static function getMediaPath()
-    {
-        return self::MEDIA_PATH;
     }
 }
